backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.1
python-multipart==0.0.6
chromadb==0.4.22
numpy<2.0
openai==1.6.1
google-generativeai==0.3.2
google-search-results
PyMuPDF==1.23.14
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1